Inline computeDose into drugs module
